feat: rename dir
