+++ golang/bitwise_op/bitwise_and.c
