+++ golang/bitwise_op/bitwise_negation_test.go
@@ -1,4 +1,4 @@
-package bitwise
+package bitwise_op
 
 import (
 	"fmt"
