+++ golang/bitwise_op/bitwise_xor.c
