+++ golang/bitwise_op/integer_bits.go
@@ -1,4 +1,4 @@
-package bitwise
+package bitwise_op
 
 import (
 	"fmt"
